requirements.txt
pyserial
pyaudio
numpy
pandas
scipy
mido
python-rtmidi